lib/src/main/java/com/thenewboston/common/dto/AccountDTO.kt
package com.thenewboston.common.dto

import com.thenewboston.data.dto.bankapi.DateSerializer
import kotlinx.datetime.LocalDateTime
import kotlinx.serialization.SerialName
import kotlinx.serialization.Serializable

@Serializable
data class AccountDTO(
    @SerialName("id")
    val id: String,

    @Serializable(with = DateSerializer::class)
    @SerialName("created_date")
    val createdDate: LocalDateTime,

    @Serializable(with = DateSerializer::class)
    @SerialName("modified_date")
    val modifiedDate: LocalDateTime?,

    @SerialName("account_number")
    val accountNumber: String,

    @SerialName("trust")
    val trust: Double
)
